requirements.txt
ollama==0.3.3
fastapi==0.115.0
uvicorn==0.30.6
instructor==1.4.3
pydantic==2.8.2
streamlit==1.38.0
psutil==6.0.0
pandas==2.2.2
httpx==0.27.2
python-dotenv==1.0.1
